feat: 백엔드 CRUD 구현

+++ Command/Command/04 Data/Entity/Mapper/ProductResultMapper.cs
@@ -1,9 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Data.Common;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Command
 {
@@ -13,7 +9,7 @@ namespace Command
         {
             data.Key.COM_CODE = reader[ProductColumns.COM_CODE].ToString();
             data.Key.PROD_CD = reader[ProductColumns.PROD_CD].ToString();
-            data.PRICE = reader.GetInt16(reader.GetOrdinal(ProductColumns.PRICE));
+            data.PRICE = reader.GetDouble(reader.GetOrdinal(ProductColumns.PRICE));
             data.PROD_NM = reader[ProductColumns.PROD_NM].ToString();
             data.WRITE_DT = (DateTime)reader[ProductColumns.WRITE_DT];
             data.IS_USED = (bool)reader[ProductColumns.IS_USED];
